devtools/client/webconsole/new-console-output/test/components/test_page-error.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for PageError component</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for PageError component</p>

<script type="text/javascript;version=1.8">
window.onload = Task.async(function* () {
  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/utils/messages");
  const { PageError } = require("devtools/client/webconsole/new-console-output/components/message-types/page-error");

  const packet = yield getPacket(null, "pageError");
  const message = prepareMessage(packet);
  const rendered = renderComponent(PageError, {message});

  const queryPath = "div.message.cm-s-mozilla span span.message-flex-body span.message-body.devtools-monospace";
  const messageBody = rendered.querySelectorAll(queryPath);
  is(messageBody.length, 1, "PageError outputs expected HTML structure");
  is(messageBody[0].textContent, testCommands.get("pageError").expectedText, "PageError outputs expected text");

  SimpleTest.finish()
});
</script>
</body>
</html>
